+++ src/app/oqapi/result/oqt-result.component.ts
@@ -12,8 +12,8 @@ import {SafeUrl} from '@angular/platform-browser';
 import {UrlHashParamsProviderService} from '../../singelton-services/url-hash-params-provider.service';
 import {IndicatorParams, Params} from '../types/types'
 import Bpolys = OhsomeApi.v1.request.Bpolys;
-import { corineLandCoverClassMapLevel2 } from '../query-form/oqt-api-query-form/thematical-accuracy-indicator/thematical-accuracy-indicator.constants';
 import { IndicatorResultComponent } from './indicator-result/indicator-result.component';
+import { categoryRegistry, thematicCategoryType, thematicAccuracyCategoryNamesForBlank } from '../query-form/oqt-api-query-form/thematic-accuracy-indicator/thematic-accuracy-indicator.constants';
 
 declare let $: any;
 
@@ -46,7 +46,11 @@ export class OqtResultComponent implements OnInit, AfterViewInit {
 
   permalink: SafeUrl;
 
-  corineLandCoverClassMapLevel2 = corineLandCoverClassMapLevel2;
+  thematicAccuracyCategories = categoryRegistry;
+
+  thematicAccuracyCategoryType = thematicCategoryType
+
+  thematicAccuracyCategoryNamesForBlank = thematicAccuracyCategoryNamesForBlank
 
   constructor() {
     this.metadata = this.oqtApiMetadataProviderService.getOqtApiMetadata();
@@ -205,11 +209,12 @@ export class OqtResultComponent implements OnInit, AfterViewInit {
 
   }
 
-  getSelectedCorineLabel(): string | null {
-    const key = String(this.formValues?.['land-cover-thematic-accuracy--corine_land_cover_class']);
-    return key && this.corineLandCoverClassMapLevel2[key]
-      ? this.corineLandCoverClassMapLevel2[key].name
-      : null;
+  getThematicAccuracyLabel(topicKey: string): string | null {
+    const key = String(this.formValues?.[topicKey + '--' + this.thematicAccuracyCategoryType[topicKey]]);
+    console.log(topicKey + '--' + this.thematicAccuracyCategoryType[topicKey])
+    return key && this.thematicAccuracyCategories[topicKey][key]
+      ? this.thematicAccuracyCategories[topicKey][key].name
+      : thematicAccuracyCategoryNamesForBlank[topicKey];
   }
 
   protected readonly Utils = Utils;