+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## current
 
 * feat: add link to indicator description page to oqapi indicators
+* feat: add option for new indicator roads-thematic-accuracy
 
 ## 1.8.1